src/requirements.txt
google-cloud-pubsub==2.12.0
kafka-python==2.0.2